+++ src/pages/Order/OrderProgress.jsx
@@ -1,6 +1,5 @@
 import { useRef, useState } from 'react';
-import { useNavigate } from 'react-router';
-import { useOutletContext } from 'react-router';
+import { useNavigate, useOutletContext } from 'react-router';
 
 import OrderButtonBox from '@/components/Order/OrderEntry/OrderButtonBox';
 import FoodNavbar from '@/components/Order/OrderProgress/FoodNavbar';
@@ -11,22 +10,11 @@ const CATEGORY_MAP = { 메인: 'main', 사이드: 'side', 음료: 'drink' };
 
 function OrderProgress() {
   const navigate = useNavigate();
-  const { boothName, foodData } = useOutletContext();
+  const { boothName, foodData, quantities, onSelect, onIncrease, onDecrease } = useOutletContext();
 
   const [activeCategory, setActiveCategory] = useState('메인');
-  const [quantities, setQuantities] = useState({});
   const scrollContainerRef = useRef(null);
   const sectionRefs = { main: useRef(null), side: useRef(null), drink: useRef(null) };
-
-  const handleSelect = (key) => setQuantities((prev) => ({ ...prev, [key]: 1 }));
-  const handleIncrease = (key) => setQuantities((prev) => ({ ...prev, [key]: prev[key] + 1 }));
-  const handleDecrease = (key) =>
-    setQuantities((prev) => {
-      const next = { ...prev };
-      if (next[key] <= 1) delete next[key];
-      else next[key] -= 1;
-      return next;
-    });
 
   const totalCount = Object.values(quantities).reduce((sum, qty) => sum + qty, 0);
   const hasSelection = totalCount > 0;
@@ -56,9 +44,9 @@ function OrderProgress() {
           sectionRefs={sectionRefs}
           activeCategory={activeCategory}
           quantities={quantities}
-          onSelect={handleSelect}
-          onIncrease={handleIncrease}
-          onDecrease={handleDecrease}
+          onSelect={onSelect}
+          onIncrease={onIncrease}
+          onDecrease={onDecrease}
         />
         <div className="h-25" />
       </div>
